+++ source/renderer/app/api/utils/patchAdaApi.js
@@ -1,6 +1,5 @@
 // @flow
 import { get, map } from 'lodash';
-import moment from 'moment';
 import { action } from 'mobx';
 import BigNumber from 'bignumber.js/bignumber';
 import AdaApi from '../api';
@@ -18,7 +17,6 @@ import type {
   NetworkInfoResponse,
 } from '../network/types';
 import type { GetNewsResponse } from '../news/types';
-import { getEpochLength } from '../../config/epochsConfig';
 
 let LOCAL_TIME_DIFFERENCE = 0;
 let SYNC_PROGRESS = null;
@@ -46,7 +44,6 @@ export default (api: AdaApi) => {
       // extract relevant data before sending to NetworkStatusStore
       const nextEpochNumber = get(next_epoch, 'epoch_number', null);
       const nextEpochStartTime = get(next_epoch, 'epoch_start_time', '');
-      const epochLength = getEpochLength();
       return {
         syncProgress: SYNC_PROGRESS !== null ? SYNC_PROGRESS : syncProgress,
         localTip: {
@@ -61,15 +58,6 @@ export default (api: AdaApi) => {
           // N+1 epoch
           epochNumber: nextEpochNumber,
           epochStart: nextEpochStartTime,
-        },
-        futureEpoch: {
-          // N+2 epoch
-          epochNumber: nextEpochNumber ? nextEpochNumber + 1 : null,
-          epochStart: nextEpochStartTime
-            ? moment(nextEpochStartTime)
-                .add(epochLength, 'seconds')
-                .toISOString()
-            : '',
         },
       };
     } catch (error) {