+++ storybook/stories/staking/DelegationSteps.stories.js
@@ -134,6 +134,7 @@ export class StakingDelegationSteps extends Component<Props, State> {
         key="DelegationStepsSuccessDialog"
         delegatedWallet={WALLETS[0]}
         delegatedStakePool={STAKE_POOLS[0]}
+        slotLength={null}
         currentLocale={this.props.locale}
         futureEpochStartTime={moment()
           .add(35, 'hour')